[update] stream in codegen c
