+++ python/heterocl/config.py
@@ -1,3 +1,2 @@
 
 init_dtype = "int32"
-init_device = "fpga_intel"
