+++ tvm/src/codegen/codegen_c.cc
@@ -8,12 +8,128 @@
 #include <iomanip>
 #include <cctype>
 #include "./codegen_c.h"
+#include "./merlinc/codeanalys_merlinc.h"
 #include "../arithmetic/compute_expr.h"
 
 namespace TVM {
 namespace codegen {
 
 using namespace ir;
+
+// collect type info for vars
+class TypeCollector final : public IRVisitor {
+  public:
+    var2nameType& top_args_;
+    TypeCollector(var2nameType& top_args)
+      : top_args_(top_args) {}
+    void Visit_(const Allocate *op) {
+      auto v = op->buffer_var.get();
+      
+      // record type and shape
+      if (top_args_.count(v)) {
+        std::vector<int> shape;
+        for (size_t i = 0; i < op->extents.size(); i++) 
+          shape.push_back(op->extents[i].as<IntImm>()->value);
+        top_args_[v] = std::make_tuple(
+                           std::get<0>(top_args_[v]),
+                           op->type, shape);
+      }
+      IRVisitor::Visit_(op);
+    }
+};
+
+// record <name, type> of vars for top func signature
+// vars include passed-in and not registered vars on host
+class StreamCollector final : public IRVisitor {
+  public:
+    StreamCollector(std::vector<const Variable*>& arg_vars,
+                    std::unordered_map<const Variable*, bool>& stream_table,
+                    std::string initial_scope)
+      : arg_vars_(arg_vars),
+        stream_table_(stream_table),
+        scope_(initial_scope) {}
+
+    // record alloc on host 
+    void Visit_(const Allocate *op) {
+      if (!switch_on) 
+        this->HandleDef(op->buffer_var.get());
+      IRVisitor::Visit_(op);
+    }
+    
+    void Visit_(const Load *op) {
+      if (!switch_on) {
+        this->HandleUse(op->buffer_var);
+      }
+      IRVisitor::Visit_(op);
+    }
+
+    // update placeholder status
+    void Visit_(const Store* op) {
+      if (!switch_on) { // count use on host
+        if (auto val = op->value.as<StreamExpr>())
+          this->HandleDef(op->buffer_var.get());
+        this->HandleUse(op->buffer_var);
+      }
+      IRVisitor::Visit_(op);
+    }
+
+    void Visit_(const StreamStmt* op) {
+      if (!switch_on) { // in host scope
+        this->HandleDef(op->buffer_var.get());
+      }
+      IRVisitor::Visit_(op);
+    }
+
+    void Visit_(const AttrStmt* op) {
+      if (op->attr_key == attr::device_scope) { 
+        if (op->value.as<StringImm>()->value != scope_)
+          switch_on = true;
+        else switch_on = false;
+      }
+      IRVisitor::Visit_(op);
+    }
+
+    // additional data saved into stream table (for streamed 
+    // data we keep the new id for arg_stream in var_idmap, 
+    // and non-streamed using the repalced arg_top_k name)
+    void HandleDef(const Variable* v) {
+      CHECK(!host_def_count_.count(v))
+          << "variable " << v->name_hint
+          << " has already been defined, the Stmt is not SSA";
+      CHECK(!host_use_count_.count(v))
+          << "variable " << v->name_hint
+          << " has been used before definition!";
+      host_use_count_[v] = 0;
+      host_def_count_[v] = 1;
+    }
+
+    void HandleUse(const Expr& v) {
+      CHECK(v.as<Variable>());
+      Var var(v.node_);
+      auto it = host_use_count_.find(var.get());
+      if (it != host_use_count_.end()) {
+        if (it->second >= 0) {
+          ++it->second;
+        }
+      } else {
+        if (!stream_table_.count(var.get())) {
+          host_undefined_.push_back(var);
+          host_use_count_[var.get()] = -1;
+        }
+      }
+    }
+
+    bool host_scope_{false};
+    Array<Var> host_undefined_;
+    std::unordered_map<const Variable*, int> host_use_count_;
+    std::unordered_map<const Variable*, int> host_def_count_;
+
+  private:
+    std::vector<const Variable*>& arg_vars_;
+    std::unordered_map<const Variable*, bool>& stream_table_;
+    std::string scope_;
+    bool switch_on{true};
+};
 
 void CodeGenC::Init(bool output_ssa) {
   print_ssa_form_ = output_ssa;
@@ -27,41 +143,45 @@ void CodeGenC::InitFuncState(LoweredFunc f) {
   CodeGenSourceBase::ClearFuncState();
 }
 
-void CodeGenC::AddFunction(LoweredFunc f) {
+void CodeGenC::AddFunction(LoweredFunc f,
+        str2tupleMap<std::string, Type> map_arg_type) {
   // clear previous generated state.
   this->InitFuncState(f);
+  map_arg_type_ = map_arg_type;
   // add to alloc buffer type.
   for (const auto & kv : f->handle_data_type) {
     RegisterHandleType(kv.first.get(), kv.second.type());
   }
 
-  // second move to generate 
+  // generate function signature 
   this->stream << "void " << f->name << "(";
   for (size_t i = 0; i < f->args.size(); ++i) {
     Var v = f->args[i];
     std::string vid = AllocVarID(v.get());
     if (i != 0) stream << ", ";
-    if (v.type().is_handle()) {
-      auto it = alloc_storage_scope_.find(v.get());
-      if (it != alloc_storage_scope_.end())
-        PrintStorageScope(it->second, stream);
-      stream << ' ';
-
-      if (handle_data_type_.count(v.get())) {
-        PrintType(handle_data_type_.at(v.get()), stream);
-      } else {
-        stream << "void";
-      }
-      stream << "*";
-
-      if (f->is_restricted && restrict_keyword_.length() != 0) {
-        stream << ' ' << restrict_keyword_;
-      }
+    // check type in the arg map
+    if (map_arg_type.find(vid) == map_arg_type.end()) {
+      LOG(WARNING) << vid << " type not found\n";
+      PrintType(v.type(), this->stream);
+      this->stream << ' ' << vid;
     } else {
-      PrintType(v.type(), stream);
+      auto arg = map_arg_type[vid];
+      PrintType(std::get<1>(arg), this->stream);
+      this->stream << "* " << std::get<0>(arg);
+      const BufferNode* buf = f->api_args[i].as<BufferNode>();
+      if (v.type().is_handle() && buf) {
+        std::vector<int> shape;
+        for (size_t i = 0; i < buf->shape.size(); i++) 
+          shape.push_back(buf->shape[i].as<IntImm>()->value);
+        arg_shapes.push_back(shape);
+        var_shape_map_[buf->data.get()] = buf->shape;
+        auto it = alloc_storage_scope_.find(v.get());
+        if (it != alloc_storage_scope_.end())
+          PrintStorageScope(it->second, stream);
+      }
     }
-    stream << ' ' << vid;
   }
+
   stream << ") {\n";
   int func_scope = this->BeginScope();
   this->PrintStmt(f->body);
@@ -95,17 +215,6 @@ std::string CodeGenC::Finish() {
          << "){\n" << device_stream.str();
   if (fpga_scope_) device << stream.str();
   else host_stream << stream.str(); 
-  // finish host call stmt
-  // if (top_data_type_.size() > 0) {
-  //   int i = 0;
-  //   for (const auto & kv : top_data_type_) {
-  //     // PrintType(kv.second, host_stream);
-  //     if (i != 0) host_stream << ", ";
-  //     host_stream << kv.first;
-  //     i++;
-  //   }
-  //   host_stream << ");\n";
-  // }
   device << "}\n";
   return decl_stream.str() + "\n{device}\n" + 
          module_stream.str() + device.str() + "\n{device}\n" + 
@@ -756,19 +865,22 @@ void CodeGenC::VisitExpr_(const GetSlice *op, std::ostream& os) { // NOLINT(*)
 }
 
 void CodeGenC::VisitExpr_(const SetBit *op, std::ostream& os) { // NOLINT(*)
-  LOG(FATAL) << "SetBit is not implemented yet";
+  LOG(FATAL) << "SetBit is not implemented yet in C";
 }
 
 void CodeGenC::VisitExpr_(const SetSlice *op, std::ostream& os) { // NOLINT(*)
-  LOG(FATAL) << "SetSlice is not implemented yet";
+  LOG(FATAL) << "SetSlice is not implemented yet in C";
 }
 
 void CodeGenC::VisitExpr_(const Quantize *op, std::ostream& os) { // NOLINT(*)
-  LOG(FATAL) << "Quantize is not yet support";
+  LOG(FATAL) << "Quantize is not yet support in C";
 }
 
 void CodeGenC::VisitExpr_(const StreamExpr *op, std::ostream& os) { // NOLINT(*)
-  LOG(FATAL) << "StreamExpr is not implemented yet";
+  auto v = op->buffer_var.get();
+  auto it = var_idmap_.find(v);
+  CHECK(it != var_idmap_.end())
+    << "variable " << v->name_hint << " not decalred";
 }
 
 void CodeGenC::VisitExpr_(const KernelExpr *op, std::ostream& os) { // NOLINT(*)
@@ -781,32 +893,56 @@ void CodeGenC::VisitExpr_(const KernelExpr *op, std::ostream& os) { // NOLINT(*)
 }
 
 void CodeGenC::VisitStmt_(const StreamStmt *op) { // NOLINT(*)
-  LOG(FATAL) << "StreamStmt is not implemented yet";
+    CHECK(!var_idmap_.count(op->buffer_var.get())); 
+    std::string vid = AllocVarID(op->buffer_var.get());
+    vid = GetVarID(op->value.as<Load>()->buffer_var.get()); 
+    PrintIndent();
+    auto load_op = op->value.as<Load>(); 
+    auto v = load_op->buffer_var.as<Variable>();
+    // placeholder args using recv name 
+    if (stream_table.count(v)) {
+      auto tuple = arg_top_vars[v];
+      arg_top_vars[v] = std::make_tuple(vid, std::get<1>(tuple),
+                                        std::get<2>(tuple));
+      stream_table[v] = true;
+    } // else: streamed externop defined in analysis
+    // PrintExpr(op->value, stream);
+    // stream << vid << ".write()\n";
 }
 
 void CodeGenC::VisitStmt_(const LetStmt* op) {
   std::string value = PrintExpr(op->value);
+  // Skip the argument retrieving assign statement
+  std::string vid = AllocVarID(op->var.get());
   if (print_ssa_form_) {
     CHECK(!var_idmap_.count(op->var.get()));
     var_idmap_[op->var.get()] = value;
   } else {
     PrintIndent();
-    if (op->var.type() == Handle() &&
-        handle_data_type_.count(op->var.get())) {
-      PrintType(handle_data_type_.at(op->var.get()), stream);
-      stream << "* "
-             << AllocVarID(op->var.get())
-             << " = (";
-      PrintType(handle_data_type_.at(op->var.get()), stream);
-      stream << "*)"  << value << ";\n";
-    } else {
+    if (op->var.type() != Handle() &&
+        value.find("TVMArray") == std::string::npos &&
+        value.find("arg") != 0) {
+      PrintIndent();
       PrintType(op->var.type(), this->stream);
       this->stream << ' '
-                   << AllocVarID(op->var.get())
+                   << vid
                    << " = " << value << ";\n";
+    // modify var idmap for passed in args
+    } else if (value.find("data") != std::string::npos ||
+               value.substr(0, 3) == "arg") {
+      auto v = op->var.get();
+      arg_vars.push_back(v);
+      stream_table[v] = false; 
+      std::string api_name = "arg" + std::to_string(arg_count);
+      auto arg = map_arg_type_[api_name];
+      // PrintType(std::get<1>(arg), arg_stream);
+      CHECK(arg_count < arg_shapes.size());
+      auto shape = arg_shapes[arg_count];
+      arg_top_vars[v] = std::make_tuple(vid, std::get<1>(arg), shape);
+      arg_count += 1;
     }
+    PrintStmt(op->body);
   }
-  PrintStmt(op->body);
 }
 
 void CodeGenC::VisitStmt_(const Allocate* op) {
@@ -841,44 +977,6 @@ void CodeGenC::VisitStmt_(const Allocate* op) {
   this->PrintStmt(op->body);
 }
 
-// record vars transferred between xcel and host
-// collect info of needed args & streamed args (types)
-class StreamCollector final : public IRVisitor {
-  public:
-    StreamCollector(std::vector<const StreamStmt*>& stream_stmt_list,
-                    std::vector<const StreamExpr*>& stream_expr_list,
-                    std::string initial_scope)
-      : stream_stmt_list_(stream_stmt_list), 
-        stream_expr_list_(stream_expr_list), 
-        scope_(initial_scope) {}
-
-    void Visit_(const StreamExpr* op) {
-      if (switch_on) {
-        stream_expr_list_.push_back(op);
-      }
-    }
-
-    void Visit_(const StreamStmt* op) {
-      if (switch_on)
-        stream_stmt_list_.push_back(op);
-    }
-
-    void Visit_(const AttrStmt* op) {
-      if (op->attr_key == attr::device_scope) { 
-        if (op->value.as<StringImm>()->value != scope_)
-          switch_on = true;
-        else switch_on = false;
-      }
-      this->Visit(op->body);
-    }
-
-  private:
-    std::vector<const StreamStmt*>& stream_stmt_list_;
-    std::vector<const StreamExpr*>& stream_expr_list_;
-    std::string scope_;
-    bool switch_on{false};
-};
-
 void CodeGenC::VisitStmt_(const AttrStmt* op) {
   if (op->attr_key == ir::attr::thread_extent) {
     IterVar iv(op->node.node_);
@@ -895,6 +993,65 @@ void CodeGenC::VisitStmt_(const AttrStmt* op) {
     const Variable* v = op->node.as<Variable>();
     CHECK(v);
     volatile_buf_.insert(v);
+  } else if (op->attr_key == ir::attr::device_scope) {
+    // print top( ... in host and enter fpga scope 
+    if (op->value.as<StringImm>()->value == "fpga" && !fpga_scope_) {
+      fpga_scope_ = true;
+      PrintIndent();
+       
+      // track the stream usage
+      StreamCollector collector(arg_vars, stream_table, "cpu");
+      collector.Visit(op->body);
+
+      // update data type and name 
+      for (auto k : collector.host_undefined_) {
+        auto v = k.get();
+        arg_vars.push_back(v);
+        stream_table[v] = true;
+        LOG(WARNING) << v->name_hint;
+        auto tuple = arg_top_vars[v];
+        arg_top_vars[v] = std::make_tuple(v->name_hint,
+                                          std::get<1>(tuple),
+                                          std::get<2>(tuple)); 
+      }
+      TypeCollector visitor(arg_top_vars);
+      visitor.Visit(op->body);
+  
+      // generte function calls 
+      stream << "top(";
+      int index = 0;
+      for (size_t i = 0; i < arg_vars.size(); i++) {
+        auto v = arg_vars[i];
+        std::string arg_name;
+        if (stream_table[v]) 
+          arg_name = std::get<0>(arg_top_vars[v]);
+        else arg_name = GetVarID(v); 
+        if (index !=0) stream << ", ";
+        stream << arg_name;
+        // print kernel func signature
+        if (index != 0) arg_stream << ", ";
+        PrintType(std::get<1>(arg_top_vars[v]), arg_stream);
+        auto shape = std::get<2>(arg_top_vars[v]);
+        arg_stream << " " << arg_name;
+        for (size_t k = 0; k < shape.size(); k++)
+          arg_stream << "[" << shape[k] << "]";
+        index++;
+      }
+      stream << ");\n";
+  
+      // switch context to device scope
+      host_stream << this->stream.str();
+      this->stream.str("");
+      this->stream.clear();
+  
+    // swtich from device to host
+    } else if (op->value.as<StringImm>()->value == "cpu" && 
+               fpga_scope_) {
+      fpga_scope_ = false;
+      device_stream << this->stream.str();
+      this->stream.str("");
+      this->stream.clear();
+    }
   }
   this->PrintStmt(op->body);
 }
