+++ tvm/src/codegen/codegen_cuda.h
@@ -10,6 +10,7 @@
 #include <tvm/packed_func_ext.h>
 #include <string>
 #include "./codegen_c.h"
+#include "./merlinc/codeanalys_merlinc.h"
 
 namespace TVM {
 namespace codegen {
@@ -18,7 +19,8 @@ class CodeGenCUDA final : public CodeGenC {
  public:
   CodeGenCUDA();
   void Init(bool output_ssa);
-  void AddFunction(LoweredFunc f);
+  void AddFunction(LoweredFunc f, 
+      str2tupleMap<std::string, Type> map_arg_type);
   // override behavior
   void VisitStmt_(const ir::For* op) final;
   void PrintStorageSync(const Call* op) final;
