+++ tvm/src/codegen/hlsc/build_hlsc.cc
@@ -46,7 +46,6 @@ std::string BuildHLSC(Array<LoweredFunc> funcs) {
     cg.AddFunction(f, map_arg_type);
   }
   std::string code = cg.Finish();
-
   LOG(WARNING) << "HLS C doesn't have runtime, return kernel code";
   return code;
 }
