+++ tvm/src/codegen/hlsc/codegen_hlsc.h
@@ -16,20 +16,20 @@ namespace TVM {
 namespace codegen {
 
 class CodeGenHLSC : public CodeGenC {
-  public:
-    void AddFunction(LoweredFunc f, str2tupleMap<std::string, Type> map_arg_type);
+ public:
+  void AddFunction(LoweredFunc f, str2tupleMap<std::string, Type> map_arg_type);
 
-    void VisitExpr_(const Min* op, std::ostream& os) override;
-    void VisitExpr_(const Max* op, std::ostream& os) override;
+  void VisitExpr_(const Min* op, std::ostream& os) override;
+  void VisitExpr_(const Max* op, std::ostream& os) override;
 
-    void VisitStmt_(const LetStmt* op) override;
-    void VisitStmt_(const IfThenElse* op) override;
-    void VisitStmt_(const Allocate* op) override;
+  void VisitStmt_(const LetStmt* op) override;
+  void VisitStmt_(const IfThenElse* op) override;
+  void VisitStmt_(const Allocate* op) override;
 
-    void GenForStmt(const For* op, std::string pragma, bool before);
-    
-  protected:
-    std::string GetBufferRef(Type t, const Variable* buffer, Expr index);
+  void GenForStmt(const For* op, std::string pragma, bool before);
+  
+ protected:
+  std::string GetBufferRef(Type t, const Variable* buffer, Expr index);
 };
 
 }  // namespace codegen
