+++ tvm/src/codegen/opencl/codegen_opencl.h
@@ -4,7 +4,6 @@
 # include <tvm/codegen.h>
 # include <tvm/packed_func_ext.h>
 # include <string>
-// # include "./codeanalys_openclc.h"
 # include "../codegen_c.h"
 
 namespace TVM{
